add feature: error code

+++ src/main/java/cn/topicstudy/jutil/basic/date/DateUtil.java
@@ -46,7 +46,7 @@ public class DateUtil {
      * 给date加上seconds秒
      * TODO fix:超过int范围
      */
-    public static Date addSeconds(Date date, Integer seconds) {
+    public static Date addSeconds(Date date, long seconds) {
         return new Date(date.getTime() + seconds * 1000);
     }
